pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biomodelml"
version = "0.1.0"
description = "Bioinformatics framework for sequence analysis using self-comparison matrices and image-based algorithms"
readme = "README.md"
requires-python = ">=3.8,<3.12"
license = {text = "MIT"}
authors = [
    {name = "BioBD Team"}
]
keywords = ["bioinformatics", "phylogenetics", "sequence-analysis", "ssim", "deep-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "pandas>=1.3.2",
    "scipy>=1.5.4",
    "numpy>=1.20.3,<2.0.0",
    "biotite>=0.30.0,<0.40.0",
    "biopython>=1.79",
    "matplotlib>=3.4.3",
    "ete3>=3.1.0,<4.0.0",
]

[project.optional-dependencies]
# Full installation with deep learning support
full = [
    "tensorflow-cpu>=2.8.0,<3.0.0",
    "opencv-python>=4.5.4.60",
    "sewar==0.4.4",
    "annoy==1.17.0",
    "ppscore>=1.2.0",
]

# GPU support (requires CUDA)
gpu = [
    "tensorflow>=2.8.0,<3.0.0",
    "opencv-python>=4.5.4.60",
    "sewar==0.4.4",
    "annoy==1.17.0",
    "ppscore>=1.2.0",
]

# Hyperparameter optimization
optimization = [
    "optuna>=2.10.0,<3.1",
]

# Development dependencies
dev = [
    "pytest>=7.1.2",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

# Documentation
docs = [
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
]

# All extras
all = [
    "biomodelml[full,optimization,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/BioBD/biomodelml"
Repository = "https://github.com/BioBD/biomodelml"
"Bug Tracker" = "https://github.com/BioBD/biomodelml/issues"

[project.scripts]
biomodelml-sanitize = "biomodelml.cli.sanitize:main"
biomodelml-matchmatrix = "biomodelml.cli.matchmatrix:main"
biomodelml-tree = "biomodelml.cli.tree_builder:main"
biomodelml-optimize = "biomodelml.cli.optimize:main"
biomodelml-clusterize = "biomodelml.cli.clusterize:main"

[tool.setuptools]
packages = ["biomodelml", "biomodelml.variants", "biomodelml.variants.deep_search", "biomodelml.cli"]

[tool.setuptools.package-data]
biomodelml = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
